Adds: webpack.config.js, webpack-dev-server, modules
