+++ public/index.html
@@ -1922,6 +1922,6 @@ Here is a card with filled data</textarea
         <button class="load-more" type="button">load more</button>
       </section>
     </main>
-    <script type="text/javascript" src="../src/main.js"></script>
+    <script type="text/javascript" src="bundle.js"></script>
   </body>
 </html>
